fix: upgrade py2neo from 3.3.1 to 3.5.28
After we upgraded edxapp from python3.5 to
python3.8, the dump_to_neo4j management
command (which we use to dump modulestore
data to into coursegraph) stopped working. Our
working hypothesis is that the py2neo version that
edx-platform pins (3.1.2) does not support python3.8.

In order upgrade to a py2neo that python3.8 supports
*and* that hasn't been deleted from Neo4j's
documentation, we must first upgrade the version of
neo4j itself that we install. By upgrading to 3.5.28,
we upgrade as far into the 3.x series as possible without
having to deal with potential breaking changes in 4.x.

According to Neo4j's docs, this should be a painless
upgrade, although it will trigger an automatic db
migration when neo4j starts up again. It should allow
us to upgrade to any py2neo version series between
4.x (the next major version) to 2021.x (the latest).

TNL-8386